API: rename icons submodule to symbols because these are not icons, causing confusion
