+++ pcdswidgets/symbols/base.py
