+++ pcdswidgets/symbols/gauges.py
